backend/requirements.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
pyjwt>=2.8.0
pydantic>=2.0.0
